@@ -10,16 +10,16 @@
         "files": [ "vendor/symfony/symfony/src/Symfony/Component/VarDumper/Resources/functions/dump.php" ]
     },
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=7.2.0",
         "doctrine/doctrine-bundle": "^1.6",
-        "doctrine/orm": "^2.5",
+        "doctrine/orm": "^2.6",
         "incenteev/composer-parameter-handler": "^2.0",
         "sensio/distribution-bundle": "^5.0.19",
         "sensio/framework-extra-bundle": "^5.0.0",
         "symfony/monolog-bundle": "^3.1.0",
         "symfony/symfony": "3.4.*",
-        "twig/twig": "^1.0||^2.0",
-        "doctrine/migrations": "~1.5",
+        "twig/twig": "^2.0",
+        "doctrine/migrations": "~1.6",
         "doctrine/doctrine-migrations-bundle": "~1.3",
         "webignition/url":">=1.0,<2.0",
         "webignition/simplytestable-user-model": ">=0.1,<1",
@@ -28,10 +28,10 @@
         "wildbit/postmark-php": "^2"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.0",
+        "phpunit/phpunit": "~7.0",
         "squizlabs/php_codesniffer": "3.*",
         "symfony/phpunit-bridge": "^3.0",
-        "mockery/mockery": "0.9",
+        "mockery/mockery": "^1.0",
         "guzzlehttp/guzzle": "~6"
     },
     "scripts": {
